task 6

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("size of a char: %d byte(s)\n", (unsigned long)sizeof(a));
-	printf("size of an int: $%d byte(s)\n", (unsigned long)sizeof(b));
-	printf("size of a long int: byte(s)\n" (unsigned long)sizeof(c));
-	printf("size of a long long int: byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of a float: byte(s)\n", (unsigned long)sizeof(e));
+	printf("size of a char: %d byte(s)\n", sizeof(a));
+	printf("size of an int: $%d byte(s)\n", sizeof(b));
+	printf("size of a long int: byte(s)\n" sizeof(c));
+	printf("size of a long long int: byte(s)\n", sizeof(d));
+	printf("size of a float: byte(s)\n", sizeof(e));
 	return (0);
 }
